Apply snapshot isolation to all SQL Server connection when it is enabled
Connections generated in BaseJdbcClient didn't set
transaction isolation even when it was enabled.
